fix: extract shared apartment data, add SEO metadata, remove unused deps
- Create data/apartments.ts as single source of truth for apartment data
- Replace hardcoded calendar URLs with environment variables
- Make bookableLink optional in IApartment interface
- Add priceFrom field to IApartment interface
- Refactor apartmentsList.tsx to re-export from shared data
- Refactor API route to import from shared data instead of duplicating
- Split apartment detail page into Server + Client components
- Add generateMetadata and generateStaticParams for SEO
- Remove unused deps: @fortawesome/*, @googlemaps/js-api-loader, @react-google-maps/api
